# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Musicfy API",
...
    "contact": {
      "name": "Musicfy Backend"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
  "paths": {
    "/v1/auth/guest": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Создать guest-сессию",
        "operationId": "guest_sign_in_v1_auth_guest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/apple": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Sign in with Apple",
        "operationId": "apple_sign_in_v1_auth_apple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Выйти (отозвать сессию)",
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "Авторизация",
          "Авторизация"
        ],
        "summary": "Текущий пользователь",
        "operationId": "me_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/analytics/events": {
      "post": {
        "tags": [
          "Аналитика"
        ],
        "summary": "Записать событие воронки",
        "operationId": "track_event_v1_analytics_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/admin/users/{user_id}/balance": {
      "get": {
        "tags": [
          "Админ"
        ],
        "summary": "Баланс пользователя",
        "operationId": "admin_balance_v1_admin_users__user_id__balance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/users/{user_id}/credits": {
      "post": {
        "tags": [
          "Админ"
        ],
        "summary": "Начислить кредиты (пак)",
        "operationId": "admin_grant_credits_v1_admin_users__user_id__credits_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCreditsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/users/{user_id}/subscription": {
      "post": {
        "tags": [
          "Админ"
        ],
        "summary": "Выдать подписку",
        "operationId": "admin_grant_subscription_v1_admin_users__user_id__subscription_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Админ"
        ],
        "summary": "Отозвать подписку",
        "operationId": "admin_revoke_subscription_v1_admin_users__user_id__subscription_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminBalanceResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "balances": {
            "items": {
              "$ref": "#/components/schemas/AdminCategoryBalance"
            },
            "type": "array",
            "title": "Balances"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "balances"
        ],
        "title": "AdminBalanceResponse"
      },
      "AdminCategoryBalance": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subscriptionRemaining": {
            "type": "integer",
            "title": "Subscriptionremaining"
          },
          "subscriptionGranted": {
            "type": "integer",
            "title": "Subscriptiongranted"
          },
          "purchasedAvailable": {
            "type": "integer",
            "title": "Purchasedavailable"
          }
        },
        "type": "object",
        "required": [
          "category",
          "subscriptionRemaining",
          "subscriptionGranted",
          "purchasedAvailable"
        ],
        "title": "AdminCategoryBalance"
      },
      "AppleSignInRequest": {
        "properties": {
          "identityToken": {
            "type": "string",
            "minLength": 1,
            "title": "Identitytoken",
            "description": "identityToken из ASAuthorizationAppleIDCredential."
          },
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce",
            "description": "Nonce, переданный в запросе Apple (если был)."
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          }
        },
        "type": "object",
        "required": [
          "identityToken"
        ],
        "title": "AppleSignInRequest"
      },
...
      "GrantCreditsRequest": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "description": "song | cover | video"
          },
          "amount": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Amount",
            "description": "Сколько кредитов начислить."
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Причина (для журнала)."
          }
        },
        "type": "object",
        "required": [
          "category",
          "amount"
        ],
        "title": "GrantCreditsRequest",
        "description": "Начислить покупные (non-expiring) кредиты в указанную категорию."
      },
      "GrantSubscriptionRequest": {
        "properties": {
          "song": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Song",
            "description": "Лимит генераций песен на период.",
            "default": 0
          },
          "cover": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Cover",
            "description": "Лимит каверов на период.",
            "default": 0
          },
          "video": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Video",
            "description": "Лимит видео на период.",
            "default": 0
          },
          "periodDays": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3650.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Perioddays",
            "description": "Длительность периода в днях (null = без срока).",
            "default": 30
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "title": "Label",
            "description": "Метка продукта/гранта.",
            "default": "admin_grant"
          }
        },
        "type": "object",
        "title": "GrantSubscriptionRequest",
        "description": "Активировать подписку и выдать периодные лимиты по категориям."
      },
      "GuestSignInRequest": {
        "properties": {
          "deviceId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviceid",
            "description": "Стабильный ID устройства (опционально)."
          }
        },
        "type": "object",
        "title": "GuestSignInRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "isGuest": {
            "type": "boolean",
            "title": "Isguest"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "isGuest",
          "displayName"
        ],
        "title": "MeResponse"
      },
...
      "SessionResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "isGuest": {
            "type": "boolean",
            "title": "Isguest"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Bearer-токен сессии."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "isGuest",
          "displayName",
          "token",
          "expiresAt"
        ],
        "title": "SessionResponse",
        "description": "Сессия. Используйте `token` как `Authorization: Bearer <token>`."
      },
...
      "TrackEventRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "properties": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TrackEventRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Админ-ключ (ADMIN_API_KEY)",
        "scheme": "bearer"
      }
    }
  },
...
}